fix(search): move search sub components from interactables to view
