--- components/interactables/Search/Result/ResultGlobal.less
@@ -1,27 +1,0 @@
-.search-result-pagination {
-  margin-top: @light-spacing;
-  padding: 5px 10px;
-  justify-content: center;
-  display: flex;
-  flex-wrap: wrap;
-  li {
-    display: block;
-    margin: 2px;
-    a {
-      display: block;
-      padding: 5px 10px;
-      &:extend(.font-secondary);
-      border-radius: @corner-rounding-smaller;
-    }
-    &.active {
-      a {
-        &:extend(.background-primary);
-      }
-    }
-    &.disabled {
-      a {
-        &:extend(.font-muted);
-      }
-    }
-  }
-}
