+++ components/views/chat/search/Input/Input.less
@@ -1,0 +1,198 @@
+.search-container {
+  position: relative;
+  font-size: @text-size;
+}
+.search-option-heading {
+  &:extend(.font-accent);
+}
+.search-box {
+  &:extend(.input-bg);
+  &:extend(.round-corners);
+  &.focus {
+    .search-input {
+      width: 250px;
+    }
+    width: 282px;
+  }
+  .search-input {
+    &:extend(.font-primary);
+    -webkit-user-select: text;
+    user-select: text;
+    display: inline-block;
+    position: relative;
+    padding: 0 5px;
+    width: 150px;
+    max-width: @full;
+    height: 30px;
+    line-height: 30px;
+    white-space: nowrap;
+    overflow: hidden;
+    font-size: @text-size;
+    background: transparent;
+    outline: 0;
+    z-index: @base-z-index;
+    scrollbar-width: 0;
+  }
+  display: inline-block;
+  position: relative;
+  width: 180px;
+  font-size: @text-size;
+  border: @light-border;
+  display: flex;
+  align-items: center;
+  overflow: hidden;
+  padding-left: 1.5rem;
+  padding-right: 40px;
+}
+.search-clear {
+  &:extend(.font-primary);
+  position: absolute;
+  font-size: @tiny-icon-size;
+  top: 5px;
+  right: 8px;
+  width: 34px;
+  height: 20px;
+  background: transparent;
+  border-color: transparent;
+  z-index: 2;
+  cursor: pointer;
+}
+.search-icon {
+  &:extend(.font-muted);
+  position: absolute;
+  top: @light-spacing;
+  left: 0;
+  width: 30px;
+  height: 30px;
+  text-align: center;
+  z-index: 0;
+  cursor: pointer;
+}
+.search-placeholder {
+  &:extend(.font-muted);
+  position: absolute;
+  top: 0;
+  left: 0;
+  height: 30px;
+  line-height: 30px;
+  font-size: @text-size;
+  padding: 0 5px;
+  padding-left: 1.8rem;
+}
+.hide {
+  display: none;
+}
+
+.search-options {
+  &:extend(.bordered);
+  &:extend(.background-semitransparent-light);
+  &:extend(.blur);
+  &:extend(.full-width);
+  &:extend(.round-corners);
+  padding: 3px 6px;
+  z-index: @base-z-index + 2;
+  position: absolute;
+  left: 0;
+  margin-top: @light-spacing * 1.5;
+}
+
+.search-option-group {
+  li {
+    &:hover:not(:first-child),
+    &.active {
+      &:extend(.font-secondary);
+      &:extend(.round-corners);
+      span:last-child {
+        &:extend(.font-secondary);
+      }
+      cursor: pointer;
+    }
+    span:last-child {
+      &:extend(.font-muted);
+    }
+    font-size: 16px;
+    padding: 5px 10px;
+  }
+
+  li:first-child {
+    font-size: 12px;
+  }
+  padding: 5px;
+}
+
+.search-recommends {
+  &:extend(.bordered);
+  &:extend(.background-semitransparent-light);
+  &:extend(.blur);
+  &:extend(.full-width);
+  &:extend(.round-corners);
+  padding: 3px 6px;
+  position: absolute;
+  left: 0;
+  margin-top: @light-spacing * 1.5;
+  z-index: @base-z-index + 2;
+}
+
+.search-recommend-group {
+  li {
+    &:hover:not(:first-child),
+    &.active {
+      background: @black;
+      border-radius: 2px;
+    }
+    span:last-child {
+      font-weight: 100;
+      color: grey;
+    }
+    font-size: 16px;
+    padding: 5px 10px;
+  }
+
+  li:first-child {
+    font-size: 12px;
+  }
+
+  .search-for {
+    white-space: nowrap;
+    text-overflow: ellipsis;
+    overflow: hidden;
+    max-width: @full;
+  }
+  padding: 5px;
+}
+
+.date-selection {
+  &:extend(.full-width);
+  &:extend(.round-corners);
+  &:extend(.bordered);
+  position: absolute;
+  left: 0;
+  margin-top: 4px;
+  z-index: @base-z-index + 2;
+  background: @semitransparent-lighter-gradient !important;
+  padding: 12px 6px;
+  display: flex;
+  justify-content: center;
+}
+
+@media only screen and (max-width: @mobile-breakpoint) {
+  .search-box {
+    &.focus {
+      .search-input {
+        width: 250px;
+        max-width: calc(@viewport-width - 200px) !important;
+      }
+      max-width: calc(100vw - 190px);
+    }
+    width: 150px;
+  }
+
+  .search-recommends,
+  .search-options {
+    margin: 0;
+    width: @full;
+    left: unset;
+    right: @light-spacing * 3;
+    top: @normal-spacing * 3;
+  }
+}
