+++ components/views/chat/search/Input/InputGlobal.less
@@ -1,0 +1,6 @@
+.search-query-highlight {
+  &:extend(.background-primary-gradient);
+  &:extend(.glow-primary);
+  padding: 3px @light-spacing;
+  border-radius: 30px;
+}
