+++ components/views/chat/search/Result/Result.html
@@ -1,0 +1,122 @@
+<div
+  class="search-result"
+  v-click-outside="toggle"
+  v-if="show && ((result && result.data) || loading === DataStateType.Loading)"
+>
+  <div class="search-result-header">
+    {{$t('search.result.search_results')}}
+    <InteractablesClose class="close" :action="toggle" />
+  </div>
+  <div class="search-result-group">
+    <div
+      v-for="item in groupList"
+      :class="{'group-item': true, 'active': item.type === groupBy}"
+      @click="toggleGroupBy(item.type)"
+    >
+      {{item.title}} {{item.count}}
+    </div>
+  </div>
+  <div class="search-result-filters">
+    <div class="container-filter user-options">
+      <multiselect
+        :options="userOptions"
+        :multiple="true"
+        :close-on-select="false"
+        :clear-on-select="false"
+        :preserve-search="true"
+        :placeholder="$t('search.result.users')"
+        label="name"
+        track-by="name"
+        v-model="queryOptions.friends"
+        selectLabel=""
+        deselectLabel=""
+        selectedLabel=""
+        @input="onChange"
+      >
+      </multiselect>
+    </div>
+    <div class="container-filter channel-options">
+      <multiselect
+        :options="channelOptions"
+        :multiple="true"
+        :clear-on-select="false"
+        :preserve-search="true"
+        :placeholder="$t('search.result.conversations')"
+        label="name"
+        track-by="name"
+        v-model="channels"
+        selectLabel=""
+        deselectLabel=""
+        selectedLabel=""
+      >
+      </multiselect>
+    </div>
+    <div class="container-filter date">
+      <v-date-picker v-model="date" is-dark>
+        <template v-slot="{ inputValue, inputEvents }">
+          <calendar-icon size="1x" class="date-icon" />
+          <input
+            :value="inputValue"
+            v-on="inputEvents"
+            class="date-input"
+            :placeholder="$t('search.result.select_date')"
+          />
+        </template>
+      </v-date-picker>
+    </div>
+    <div class="container-filter">
+      <button class="add-more-filters">
+        {{$t('search.result.more_filters')}}
+      </button>
+    </div>
+  </div>
+  <div
+    class="search-result-meta"
+    v-if="loading !== DataStateType.Loading && result && result.data && result.data.totalRows"
+  >
+    <div class="result-count">
+      {{result.data.totalRows}} {{$t('search.result.results')}}
+    </div>
+    <div class="result-orderby">
+      <div
+        v-for="item in orderTypeList"
+        :class="{'orderby-item': true, 'active': item.type === orderBy}"
+        @click="toggleOrderBy(item.type)"
+      >
+        {{item.title}}
+      </div>
+    </div>
+  </div>
+  <UiScroll
+    verticalScroll
+    scrollbarVisibility="scroll"
+    class="search-result-content"
+  >
+    <div class="search-result-inner">
+      <UiLoadersMessage :count="2" v-if="loading === DataStateType.Loading" />
+      <template v-else>
+        <SearchResultItem
+          v-for="item, i in result.data.list"
+          :data="item"
+          :key="i"
+        ></SearchResultItem>
+      </template>
+    </div>
+  </UiScroll>
+  <paginate
+    v-if="result && result.data && result.data.totalRows"
+    :page-count="Math.ceil(result.data.totalRows/result.data.perPage)"
+    v-model="page"
+    prev-text="«"
+    next-text="»"
+    container-class="search-result-pagination"
+    :click-handler="handleClickPaginate"
+  />
+  <div class="search-result-history">
+    <InteractablesChip
+      v-for="(item, i) in givenQueryItems"
+      :key="i"
+      :text="item"
+    />
+  </div>
+</div>
