+++ components/views/chat/search/Result/Result.vue
